addet users online

+++ www/cms2/admin/includes/view_all_posts.php
@@ -17,6 +17,33 @@
                         $update_to_published_status = mysqli_query($connect,$query);
                         confirm($update_to_published_status);
                         break;
+
+                    case 'clone':
+                        $query = "SELECT * FROM posts WHERE post_id= '{$postValueId}'";
+                        $select_posts=
+                            mysqli_query($connect, $query);
+                        confirm($select_posts);
+                        while ($row = mysqli_fetch_assoc($select_posts)) {
+                            $post_id = $row['post_id'];
+                            $post_author = $row['post_author'];
+                            $post_title = $row['post_title'];
+                            $post_category_id = $row['post_category_id'];
+                            $post_content = $row['post_content'];
+                            $post_status = $row['post_status'];
+                            $post_image = $row['post_image'];
+                            $post_tags = $row['post_tags'];
+                            $post_comments_count = $row['post_comments_count'];
+                            $post_date = $row['post_date'];
+                            $post_view_count = $row['post_view_count'];
+                        }
+                        $query = "INSERT INTO `posts`( `post_category_id`, `post_title`, `post_author`, `post_image`,
+                                       `post_content`, `post_date`, `post_tags`, `post_status`)
+                          VALUES ({$post_category_id},'{$post_title}','{$post_author}','{$post_image}','{$post_content}',
+                          '{$post_date}','{$post_tags}','{$post_status}')";
+                                    $create_post_query = mysqli_query($connect, $query);
+                                    confirm($create_post_query);
+                        break;
+
                     case 'delete':
                         $query = " DELETE FROM posts  WHERE post_id = '{$postValueId}' ";
                         $update_to_published_status = mysqli_query($connect,$query);
@@ -38,6 +65,7 @@
                             <option value="">Select Options</option>
                             <option value="published">Published</option>
                             <option value="draft">Draft</option>
+                            <option value="clone">Clone</option>
                             <option value="delete">Delete</option>
                         </select>
                     </div>
@@ -63,6 +91,7 @@
         <td>Tags</td>
         <td>Comments</td>
         <td>Status</td>
+        <td>View Count</td>
         <td>Date</td>
         <td>View</td>
         <td>Edit</td>
@@ -73,7 +102,7 @@
     <tbody>
     <?php
 
-    $query = "SELECT * FROM posts";
+    $query = "SELECT * FROM posts ORDER BY post_id DESC";
     $select_posts=
         mysqli_query($connect, $query);
     while ($row = mysqli_fetch_assoc($select_posts)) {
@@ -86,6 +115,7 @@
         $post_tags = $row['post_tags'];
         $post_comments_count = $row['post_comments_count'];
         $post_date = $row['post_date'];
+        $post_view_count = $row['post_views_count'];
         echo "<tr>";?>
         <td><input class='checkBoxes' type='checkbox' name='checkBoxArray[]'
             value='<?php echo $post_id;?>'></td>
@@ -111,10 +141,11 @@
         echo "<td>{$post_tags}</td>";
         echo "<td>{$post_comments_count}</td>";
         echo "<td>{$post_status}</td>";
+        echo "<td><a href='posts.php?reset=$post_id'>{$post_view_count}</a></td>";
         echo "<td>{$post_date}</td>";
         echo "<td><a href='../post.php?p_id={$post_id}'>View</a></td>";
         echo "<td><a href='posts.php?source=edit_post&p_id={$post_id}'>Edit</a></td>";
-        echo "<td><a href='posts.php?delete=$post_id'>Delete</a></td>";
+        echo "<td><a onclick=\" javascript: return confirm('Are you shoure you whan to delete')\" href='posts.php?delete=$post_id'>Delete</a></td>";
         echo "</tr>";
 
     }
@@ -128,6 +159,13 @@ if (isset($_GET['delete'])){
     $the_post_id = $_GET['delete'];
     $query = "DELETE FROM posts WHERE post_id = {$the_post_id}";
     $delete_query =mysqli_query($connect,$query);
+    header('Location: posts.php');
+}
+
+if (isset($_GET['reset'])){
+    $the_post_id = $_GET['reset'];
+    $query = "UPDATE posts SET post_views_count =0 WHERE post_id = {$the_post_id}";
+    $reset_query =mysqli_query($connect,$query);
     header('Location: posts.php');
 }
